Create 84_largestRectangle
Find left boundary and right boundary for each index of height.
To find the boundary similar to water tank which is the next smaller element, we can use monotonic stack to help we initialize nextSmaller and previousSmaller arrays. We are constructing an increasing order monotonic stack. Whenever we find some element smaller than the top element of the stack, which breaks the increasing order, we will pop elements out of the stack util we make sure for every elements in the stack, next smaller is stored in stack just close by to it.